chore: SelectProps 타입 분리
