+++ src/types/Register/RegisterType.tsx
@@ -1,3 +1,11 @@
 export interface OptionsProps {
   onSelectOption: (option: string) => void;
 }
+
+export interface SelectProps {
+  inputValue: string;
+  isOpen: boolean;
+  selectedTag: string;
+  onToggleDropdown: (e: React.MouseEvent<HTMLDivElement> | null) => void;
+  onTagChange: (value: string) => void;
+}
